Address PR review feedback on org create command
- Move OrgInfo and CreateOrg from projects.go into a new orgs.go
- Hardcode vcs_type as "circleci" in CreateOrg (standalone only)
- Hide the org command from help output
